nexus-api/requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.30.0
networkx>=3.0
openai>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
